+++ misc/src/hkim0331/misc.clj
@@ -3,6 +3,7 @@
    [clojure.string :as str])
   (:import java.util.Base64))
 
+
 ;;; string abbreviation
 (defn abbrev
   "abbreviate string s. default 8 characters.
@@ -13,17 +14,6 @@
    (let [re (re-pattern (format "^(.{%d}).*" n))]
      (str/replace s re (str "$1" "...")))))
 
-;; conversion between `char` and `int`
-;;
-;; no use!
-;;
-;; (defn char->int
-;;   [c]
-;;   (int c))
-;;
-;; (defn int->char
-;;   [n]
-;;   (char n))
 
 ;;; Base 64
 (defn ->base64
@@ -39,6 +29,7 @@
        (map char)
        (apply str)))
 
+
 ;; compress/expand
 (defn compress
   "compress message
@@ -51,19 +42,35 @@
        (map #(str (first %) (count %)))
        (apply str)))
 
+(defn digit?
+  [c]
+  (seq (filter (partial = c) "0123456789")))
+
+(defn small-alpha?
+  [c]
+  (seq (filter (partial = c) "abcdefghijklmnopqrstuvwxyz")))
+
+(defn large-alpha?
+  [c]
+  (seq (filter (partial = c) "ABCDEFGHIJKLMNOPQRSTUVWXYZ")))
+
+(defn alpha?
+  [c]
+  (or (small-alpha? c) (large-alpha? c)))
+
 (defn- expand-1
-  [[[c] [n]]]
-  (->> (repeat (char->int n) c)
-      (apply str)))
+  [[[c] coll]]
+  (let [n (-> (apply str coll) Integer/parseInt)]
+    (->> (repeat n c) (apply str))))
 
 (defn expand
   "retrieve original message from compressed one"
   [s]
-  (->> s
-       (partition-by digit?)
+  (->> (partition-by alpha? s)
        (partition 2)
        (map expand-1)
        (apply str)))
+
 
 ;;; qsort
 (defn- smaller
@@ -99,3 +106,32 @@
        (partition 2 1)
        (map #(<= (first %) (second %)))
        (every? true?)))
+
+
+;;; fold
+;; https://stackoverflow.com/questions/16800255/how-do-we-do-both-left-and-right-folds-in-clojure
+(defn fold-l
+  [f val coll]
+  (if (empty? coll)
+    val
+    (fold-l f (f val (first coll)) (rest coll))))
+
+(defn fold-r
+  [f val coll]
+  (if (empty? coll)
+    val
+    (f (fold-r f val (rest coll)) (first coll))))
+
+(comment
+  (fold-l + 0 (range 10))
+  (fold-r + 0 (range 10)))
+
+
+;;; inject
+(defn- inject-aux [coll ret]
+  (if (empty? coll)
+    ret
+    (inject-aux (rest coll) (conj ret (* (last ret) (first coll))))))
+
+(defn inject [coll]
+  (reduce + (inject-aux coll [1])))