+++ CHANGELOG.md
@@ -1,11 +1,24 @@
 # clojure-libs
 
-## [Unreleased]
+## Unreleased
 - Java interface is one of my weak points.
+### Bug
+ ```
+(let [e (hj/now-in-epoch)
+      e2 (hj/datetime->epoch (hj/epoch->datetime e)
+     (- e e2))
+```
 
-## 0.2.1-SNAPSHOT
+## [0.2.2] - 2022-09-07
+### Fixed
+- fix typo in javatime/javatime.clj
+- misc/expand
+
+## [0.2.1] - 2022-09-07
 ### Changed
 - move `sort` into `misc`.
+- instead of epoch, use second for function names. for example,
+  `now-in-epoch` was renamed as `now-in-second`.
 
 ## [0.2.0] - 2022-09-06
 ### Added
@@ -14,6 +27,8 @@
 ## 0.1.0 - 2022-09-06
 - initialize repository
 
-[Unreleased]:https://github.com/hkim0331/clojure-libs/compare/0.2.0...HEAD
+
+[0.2.2]:https://github.com/hkim0331/clojure-libs/compare/0.2.2...0.2.1
+[0.2.1]:https://github.com/hkim0331/clojure-libs/compare/0.2.1...0.2.0
 [0.2.0]: https://github.com/hkim0331/clojure-libs/compare/0.1.0...0.2.0
 [0.1.1]: https://github.com/hkim0331/clojure-libs/compare/0.1.0...0.1.1